Reorganized according to ecma and components
