+++ bookmarks.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Quiz App Bookmarks</title>
     <link rel="stylesheet" href="./styles/index.css" />
-    <script defer src="/javascript.js"></script>
+    <script type="module" defer src="./index.js"></script>
   </head>
   <body>
     <div class="app-grid h-auto">
