+++ js/answer.js
@@ -1,11 +1,4 @@
-export default function Card(cardElement) {
-  const BookmarkElement = document.querySelector('[data-js="bookmark"]');
-  const ButtonToggle = document.querySelector('[data-js="button-toggle" ]');
-
-  ButtonToggle.addEventListener('click', () => {
-    BookmarkElement.classList.toggle('card__bookmark--active');
-  });
-
+export function answers() {
   const answerButton = document.querySelector('[data-js="show-answer"]');
   const hideButton = document.querySelector('[data-js="hide-answer"]');
