+++ profile.html
@@ -6,6 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Quiz App Profile</title>
     <link rel="stylesheet" href="./styles/index.css" />
+    <script type="module" defer src="./index.js"></script>
   </head>
   <body>
     <div class="app-grid h-auto">
